+++ GANDLF/models/unet_multilayer.py
@@ -12,37 +12,7 @@ from GANDLF.models.seg_modules.out_conv import out_conv
 from .modelBase import ModelBase
 import sys
 import numpy as np
-
-
-def checkPatchDimensions(patch_size, numlay):
-    if isinstance(patch_size, int):
-        patch_size_to_check = np.array(patch_size)
-    else:
-        patch_size_to_check = patch_size
-    # for 2D, don't check divisibility of last dimension
-    if patch_size_to_check[-1] == 1:
-        patch_size_to_check = patch_size_to_check[:-1]
-
-    if all(
-        [x >= 2 ** (numlay + 1) and x % 2**numlay == 0 for x in patch_size_to_check]
-    ):
-        return numlay
-    else:
-        # base2 = np.floor(np.log2(patch_size_to_check))
-        base2 = np.array([getBase2(x) for x in patch_size_to_check])
-        remain = patch_size_to_check / 2**base2  # check that at least 1
-
-        layers = np.where(remain == 1, base2 - 1, base2)
-        return int(np.min(layers))
-
-
-def getBase2(num):
-    base = 0
-    while num % 2 == 0:
-        num = num / 2
-        base = base + 1
-    return base
-
+from GANDLF.utils.generic import checkPatchDimensions
 
 class unet(ModelBase):
     """
