+++ GANDLF/utils/generic.py
@@ -108,3 +108,44 @@ def version_check(version_from_config, version_to_check):
         )
 
     return True
+
+
+def checkPatchDimensions(patch_size, numlay):
+    """
+    This function checks the divisibility of a numpy array or integer for architectural integrity
+
+    Args:
+        patch_size (numpy.array): The patch size for checking.
+        number (int, optional): The number to check divisibility for. Defaults to 16.
+
+    Returns:
+        int: Largest multiple of 2 (less than or equal to numlay) that each element of patch size is divisible by to yield int >= 2
+    """
+    if isinstance(patch_size, int):
+        patch_size_to_check = np.array(patch_size)
+    else:
+        patch_size_to_check = patch_size
+    # for 2D, don't check divisibility of last dimension
+    if patch_size_to_check[-1] == 1:
+        patch_size_to_check = patch_size_to_check[:-1]
+
+    if all(
+        [x >= 2 ** (numlay + 1) and x % 2**numlay == 0 for x in patch_size_to_check]
+    ):
+        return numlay
+    else:
+        # base2 = np.floor(np.log2(patch_size_to_check))
+        base2 = np.array([getBase2(x) for x in patch_size_to_check])
+        remain = patch_size_to_check / 2**base2  # check that at least 1
+
+        layers = np.where(remain == 1, base2 - 1, base2)
+        return int(np.min(layers))
+
+
+def getBase2(num):
+    # helper for checkPatchDimensions (returns the largest multiple of 2 that num is evenly divisible by)
+    base = 0
+    while num % 2 == 0:
+        num = num / 2
+        base = base + 1
+    return base
